requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
djangorestframework==3.15.2
idna==3.7
pyTelegramBotAPI==4.21.0
requests==2.32.3
setuptools==70.3.0
sqlparse==0.5.0
urllib3==2.2.2
wheel==0.43.0
